+++ app/components/game/index.ts
@@ -3,3 +3,4 @@ export * from "./stats-card";
 export * from "./inline-editable-content";
 export * from "./log-entry-bubble";
 export * from "./log-block-bubble";
+export * from "./error-tooltip";